groups fix.

+++ socialService/src/services/groups.service.ts
@@ -4,12 +4,12 @@ import { TYPES } from './../ioc-container/types';
 import IGroupsService from '../interfaces/services/groups-service.interface';
 import IGroupsRepository from '../interfaces/repositories/group-repository.interafce';
 import IGroup from '../interfaces/models/group.interface';
+import { group } from 'console';
 @injectable()
 export default class GroupsService implements IGroupsService {
   constructor(@inject(TYPES.IGroupsRepository) private groupsRepo: IGroupsRepository) {}
 
   async addNewGroup(creatorId: string, name: string): Promise<IGroup> {
-    console.log(name);
     if (await this.groupsRepo.checkIfNameValid(name)) {
       const newGroup = await this.groupsRepo.addGroup({ creatorId: creatorId, name: name });
       newGroup.id && (await this.groupsRepo.addFriendToGroup(newGroup.id, creatorId));
@@ -17,26 +17,32 @@ export default class GroupsService implements IGroupsService {
     } else throw new Error('name is already taken');
   }
 
-  async removeGroup(groupId: string): Promise<boolean> {
-    return (await this.groupsRepo.removeGroup(groupId)) > 0;
+  async removeGroup(groupId: string, userId: string): Promise<boolean> {
+    this.validateCreator(groupId, userId);
+    if (this.groupsRepo.checkIfUserIsCreator(groupId, userId)) return (await this.groupsRepo.removeGroup(groupId)) > 0;
   }
-  async getGroup(groupId: string): Promise<IGroup> {
+  async getGroup(groupId: string, userId: string): Promise<IGroup> {
+    this.validateCreator(groupId, userId);
     return this.groupsRepo.getGroup(groupId);
   }
   getUsersGroups(userId: string): Promise<IGroup[]> {
     return this.groupsRepo.getUsersGroups(userId);
   }
-  addUserToGroup(groupId: string, userId: string): Promise<boolean> {
-    return this.groupsRepo.addFriendToGroup(groupId, userId);
+  addUserToGroup(groupId: string, userId: string, friendId: string): Promise<boolean> {
+    this.validateCreator(groupId, userId);
+    return this.groupsRepo.addFriendToGroup(groupId, friendId);
   }
-  removeUserFromGroup(groupId: string, userId: string): Promise<boolean> {
+  removeUserFromGroup(groupId: string, userId: string, friendId: string): Promise<boolean> {
+    this.validateCreator(groupId, userId);
     return this.groupsRepo.removeFriendFromGroup(groupId, userId);
   }
-  async updateGroupName(groupId: string, name: string, creatorId: string): Promise<IGroup> {
+  async updateGroupName(groupId: string, name: string, userId: string): Promise<IGroup> {
+    this.validateCreator(groupId, userId);
     if (this.groupsRepo.checkIfNameValid(name)) {
-      const groupFromDb = await this.groupsRepo.getGroup(groupId);
-      if (groupFromDb?.creatorId === creatorId) return this.groupsRepo.updateGroup(groupId, { creatorId, name });
-      else throw new Error('group Didnt Change user is not creator');
-    } else throw new Error('name already taken');
+      return this.groupsRepo.updateGroup(groupId, { creatorId: userId, name });
+    } else throw new Error('group Didnt Change user is not creator');
+  }
+  private async validateCreator(groupId: string, userId: string) {
+    if (!(await this.groupsRepo.checkIfUserIsCreator(groupId, userId))) throw new Error('name already taken');
   }
 }
